# wikiwatch-backend/WikiWatch-openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "WikiWatch API",
    "description": "Instructor-provided backend for frontend training. One shared team. Real database, role permissions, review transitions and plain-text diff comments. Public Wikimedia reads stay in the browser. Login, copy access_token, then use Authorize. All JSON success responses use {success, data}.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/wikiwatch-service/v1/preferences/me": {
      "get": {
        "tags": [
          "View preferences"
        ],
        "summary": "Read saved board order and viewed edits",
        "operationId": "read_wikiwatch_service_v1_preferences_me_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse_PreferenceResponse_"
                }
              }
            }
          },
          "401": {
            "description": "Session expired or invalid",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Role or ownership denied",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Conflict: reload before retry",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Invalid request or transition reason"
          },
          "429": {
            "description": "Login rate limit reached"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "put": {
        "tags": [
          "View preferences"
        ],
        "summary": "Save board order and viewed edits",
        "description": "Reordering columns changes presentation only. It does not change review states. A stale version returns 409.",
        "operationId": "save_wikiwatch_service_v1_preferences_me_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PreferenceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse_PreferenceResponse_"
                }
              }
            }
          },
          "401": {
            "description": "Session expired or invalid",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Role or ownership denied",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Conflict: reload before retry",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Invalid request or transition reason"
          },
          "429": {
            "description": "Login rate limit reached"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorResponse": {
        "properties": {
          "error": {
            "type": "string",
            "title": "Error"
          }
        },
        "type": "object",
        "required": [
          "error"
        ],
        "title": "ErrorResponse"
      },
...
      "PreferenceRequest": {
        "properties": {
          "version": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Version",
            "description": "Version from the last read. A stale version returns 409."
          },
          "board_order": {
            "items": {
              "type": "string",
              "enum": [
                "unclaimed",
                "claimed",
                "flagged",
                "returned",
                "reviewed"
              ]
            },
            "type": "array",
            "title": "Board Order"
          },
          "viewed_edit_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 1000,
            "title": "Viewed Edit Ids"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "version"
        ],
        "title": "PreferenceRequest"
      },
      "PreferenceResponse": {
        "properties": {
          "version": {
            "type": "integer",
            "title": "Version"
          },
          "board_order": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Board Order"
          },
          "viewed_edit_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Viewed Edit Ids"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "version",
          "board_order",
          "viewed_edit_ids"
        ],
        "title": "PreferenceResponse"
      },
...
      "SuccessResponse_PreferenceResponse_": {
        "properties": {
          "success": {
            "type": "boolean",
            "const": true,
            "title": "Success",
            "default": true
          },
          "data": {
            "$ref": "#/components/schemas/PreferenceResponse"
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "SuccessResponse[PreferenceResponse]"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "description": "Paste the access_token returned by login. Do not use the refresh token.",
        "scheme": "bearer"
      }
    }
  }
}